CMakeLists: Fix possible race condition
There seems to be a race condition that can make the build process fail. It seems to be related to mupen64plus-core and happens randomly, especially on an old and slow MacBook. Assuming it's the actual cause, it might be happening whenever the core is built with and without the New Dynarec in parallel, making its object files clash with each other as their names are the exact same.